[FEATURE] authorization

+++ src/modules/catalog/infrastructure/http/routes/sku-attribute-value-routes.ts
@@ -5,6 +5,7 @@ import { ISkuAttributeValueRepository } from '@modules/catalog/application/inter
 import { TypeOrmSkuAttributeValueRepository } from '../../persistence/TypeOrmSkuAttributeValueRepository';
 import { validate } from '@shared/middlewares/validator';
 import { attributeSkuIdValueParamUuidSchema, attributeValueIdParamUuidSchema, createAtributeValueSchema, uuidParamUuidSchema } from '../validation-schemas/sku-attribute-value-schemas';
+import { authorize, UserRole } from '@shared/middlewares/authorization-middleware';
 
 export const skuAttributeRoutes = Router();
 
@@ -14,6 +15,7 @@ const controller = new SkuAttributeValueController(service);
 
 skuAttributeRoutes.post(
     '/',
+    authorize([UserRole.ADMIN]),
     validate(createAtributeValueSchema),
     (req, res) => controller.create(req, res)
 );
@@ -26,12 +28,14 @@ skuAttributeRoutes.get(
 
 skuAttributeRoutes.put(
     '/:id',
+    authorize([UserRole.ADMIN]),
     validate(attributeValueIdParamUuidSchema),
     (req, res) => controller.update(req, res)
 );
 
 skuAttributeRoutes.delete(
     '/:id',
+    authorize([UserRole.ADMIN]),
     validate(uuidParamUuidSchema),
     (req, res) => controller.delete(req, res)
 );
